requirements.txt
djangorestframework==3.9.0
Django==2.1.3

